divide role
