+++ midi/src/main/java/com/yapp/midi/service/MidiSaveServiceImpl.java
@@ -15,5 +15,5 @@ import com.yapp.midi.util.Encrypt;
 @Service("MidiSaveService")
 public class MidiSaveServiceImpl implements MidiSaveService{
 	
-
+	
 }
